Move the engine debugger to the TypeScript adapter layer
